[alarms] Support for read-only calendar alarms, add type property
If the Maemo::Timed::Event event contains the attributes startDate
and endDate which are valid ISO 8601 extended format dates, then
interpret the event as a calendar alarm. Calendar alarms are
read only, they cannot be added through this plugin, only read.

The startDate QML property indicates when the event starts,
endDate indicates when the event ends. The uid QML property is
the unique identifier of a calendar event.

The allDay QML property simply indicates whether the startDate and
endDate times are both 00:00, in which case the calendar alarm
is interpreted as an all day event.

The calendar specific properties are read from the
Maemo::Timed::Event attributes which have the same names as the
QML properties described above.

The type QML property is one of the enumeration values
Alarm::Calendar, Alarm::Countdown, and Alarm::Clock,
the property indicates the type of the alarm.